% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_app.R
\name{build_app}
\alias{build_app}
\title{Build ActiveForms as app via Electron}
\usage{
build_app(build_path)
}
\arguments{
\item{build_path}{Directory to build the app to}
}
\description{
Build ActiveForms as app via Electron
}
